pyproject.toml
[project]
name = "registry-voidspace-ru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.5",
    "aiomysql>=0.3.2",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy[asyncio]>=2.0.49",
]
